refactor(frontend): ストーリー共通ユーティリティ導入とスケール取得の共通化 (CHASE-102)
- parts/story-utils.tsを追加し、selectText/monoスタイルを共通化
- transitions/spacing/sizing/borders/shadows/typographyで共通ユーティリティを利用
- DesignTokenHelperのgetSpacingScale/getSizeScaleの重複ロジックをgetScale(category)に集約
- lint/typecheckパスを確認
